+++ tests/StaticAnalyserTest.php
@@ -23,7 +23,7 @@ class StaticAnalyserTest extends SwaggerTestCase
     {
         $analyser = new StaticAnalyser();
         $analysis = $analyser->fromFile(__DIR__ . '/Fixtures/routes.php');
-        $this->assertCount(16, $analysis->annotations);
+        $this->assertCount(18, $analysis->annotations);
     }
     
     public function testTrait()